change int variable to double

+++ lib/models/weather_response.dart
@@ -25,7 +25,7 @@ class Location {
 }
 
 class Current {
-  int? tempC;
+  double? tempC;
   int? humidity;
   Condition? condition;
   Current({this.tempC, this.condition, this.humidity});
